WIP: rewrite resources center admin with final form
